requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.1
boto3==1.26.85 
